finished metadata updates for leaderboard
